inital import of old calendar gem
This appears to be an orphaned gem, which provides a brilliant (if somewhat old-style) interface to dealing with dates in a generic way on many calendars.
If anyone knows the original author, I'm happy to hand over my repo to him (if he cares to maintain it) -- for my part, I'm going to perform a hostile takeover
and begin maintainence, starting with the removal of the C code.

May God have mercy on our souls.